rename intent and rewrite it to cycle through favorites when buttons are pressed
